requirements.txt
Flask==2.1.2
Flask-Cors==3.0.10
flask-restx==0.5.1
python-dotenv==0.19.0
kthread==0.2.2
opencv-python==4.10.0.84
Pillow
mongoengine==0.23.1
pymilvus==2.4.6
minio==7.1.0
torch==1.9.0
numpy==1.24.4
torchvision==0.10.0
Unidecode==1.2.0
lap==0.4.0
filterpy
numba
ffmpeg-python
moviepy
scikit-image
scikit-learn
pandas
pyyaml
matplotlib
seaborn
ffmpeg-python
imutils
vidgear
requests
sentry-sdk==0.10.2
msgpack-python
schedule
whatimage
pyheif
pillow_heif
sentry-sdk==0.10.2
pika
ultralytics
posix_ipc
werkzeug==2.1.2
protobuf==3.20.3
imageio==2.34.2
